requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
pypdf==5.6.1
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.2.28
langchain-ollama==1.1.0
langchain-pinecone==0.2.13
langchain-openai==1.1.12
pinecone==7.3.0
sentence-transformers==4.1.0
requests==2.33.1